classes in separate files fix (old) Qt headers includes
