+++ src/Network/HTTPWindow.h
@@ -13,18 +13,13 @@
 
 #include <string>
 
-#include <QtGui>
-#if QT_VERSION >= 0x050000
-#include <QtWidgets>
-#endif
-
-#include <QtNetwork>
+#include <QMainWindow>
 
 #include "faust/dsp/dsp.h"
 
 using namespace std;
 
-class HTTPWindow : public QMainWindow{
+class HTTPWindow : public QMainWindow {
     
     private:
